[Refactor] refactor delta lake metastore interface
